LayoutTests/js/timezone-format-en-expected.txt

PASS date1.toString() is "Fri Apr 29 2022 13:10:43 GMT-0700 (Pacific Daylight Time)"
PASS date1.toTimeString() is "13:10:43 GMT-0700 (Pacific Daylight Time)"
PASS successfullyParsed is true

TEST COMPLETE
